Remove references to monthly_billing table from api
