+++ app/celery/scheduled_tasks.py
@@ -19,7 +19,6 @@ from app.celery.service_callback_tasks import (
 )
 from app.celery.tasks import process_job
 from app.config import QueueNames, TaskNames
-from app.dao.date_util import get_month_start_and_end_date_in_utc
 from app.dao.inbound_sms_dao import delete_inbound_sms_created_more_than_a_week_ago
 from app.dao.invited_org_user_dao import delete_org_invitations_created_more_than_two_days_ago
 from app.dao.invited_user_dao import delete_invitations_created_more_than_two_days_ago
@@ -29,10 +28,6 @@ from app.dao.jobs_dao import (
     dao_get_jobs_older_than_limited_by
 )
 from app.dao.jobs_dao import dao_update_job
-from app.dao.monthly_billing_dao import (
-    get_service_ids_that_need_billing_populated,
-    create_or_update_monthly_billing
-)
 from app.dao.notifications_dao import (
     dao_timeout_notifications,
     is_delivery_slow_for_provider,
@@ -67,9 +62,6 @@ from app.models import (
 )
 from app.notifications.process_notifications import send_notification_to_queue
 from app.performance_platform import total_sent_notifications, processing_time
-from app.utils import (
-    convert_utc_to_bst
-)
 from app.v2.errors import JobIncompleteError
